+++ RIS.Collections/Chunked/SyncArrayLog2/SyncChunkedArrayLog2L.cs
@@ -128,7 +128,7 @@ namespace RIS.Collections.Chunked
 
         }
         public SyncChunkedArrayLog2L(long length)
-            : this(length, Environment.GCLOHThresholdSize / Environment.GetSize<T>())
+            : this(length, Environment.GCLOHThresholdSize / (uint)Environment.GetSize<T>())
         {
 
         }
